serverless_ctf/ctf_app/requirements.txt
boto3==1.9.4
pyjwt==1.6.4
python-docx==0.8.4